+++ src/assets/icons/index.ts
@@ -1,8 +1,8 @@
 export { Bookmark } from './bookmark'
+export { Create } from './create'
 export { Home } from './home'
-export { LogOut } from './log-out'
+export { LogOut } from './logOut'
 export { Message } from './message'
 export { Person } from './person'
-export { PlusSquare } from './plus-square'
 export { Search } from './search'
 export { TrendingUp } from './trending-up'
