--- src/assets/icons/plus-square.tsx
@@ -1,64 +1,0 @@
-export const PlusSquare = () => {
-  return (
-    <svg
-      fill={'none'}
-      height={'24.000000'}
-      viewBox={'0 0 24 24'}
-      width={'24.000000'}
-      xmlns={'http://www.w3.org/2000/svg'}
-    >
-      <desc>Created with Pixso.</desc>
-      <defs>
-        <clipPath id={'clip301_4459'}>
-          <rect
-            fill={'white'}
-            fillOpacity={'0'}
-            height={'23.000000'}
-            id={'plus-square-outline'}
-            rx={'-0.500000'}
-            transform={'translate(0.500000 0.500000)'}
-            width={'23.000000'}
-          />
-        </clipPath>
-      </defs>
-      <rect
-        fill={'inherit'}
-        fillOpacity={'0'}
-        height={'23.000000'}
-        id={'plus-square-outline'}
-        rx={'-0.500000'}
-        transform={'translate(0.500000 0.500000)'}
-        width={'23.000000'}
-      />
-      <g clipPath={'url(#clip301_4459)'}>
-        <g opacity={'0.000000'}>
-          <path
-            d={'M0 0L24 0L24 24L0 24L0 0Z'}
-            fill={'inherit'}
-            fillOpacity={'1.000000'}
-            fillRule={'evenodd'}
-            id={'Vector'}
-          />
-        </g>
-        <path
-          d={
-            'M18 3L6 3C5.2 3 4.44 3.31 3.87 3.87C3.31 4.44 3 5.2 3 6L3 18C3 18.79 3.31 19.55 3.87 20.12C4.44 20.68 5.2 21 6 21L18 21C18.79 21 19.55 20.68 20.12 20.12C20.68 19.55 21 18.79 21 18L21 6C21 5.2 20.68 4.44 20.12 3.87C19.55 3.31 18.79 3 18 3ZM19 18C19 18.26 18.89 18.51 18.7 18.7C18.51 18.89 18.26 19 18 19L6 19C5.73 19 5.48 18.89 5.29 18.7C5.1 18.51 5 18.26 5 18L5 6C5 5.73 5.1 5.48 5.29 5.29C5.48 5.1 5.73 5 6 5L18 5C18.26 5 18.51 5.1 18.7 5.29C18.89 5.48 19 5.73 19 6L19 18Z'
-          }
-          fill={'inherit'}
-          fillOpacity={'1.000000'}
-          fillRule={'nonzero'}
-          id={'Vector'}
-        />
-        <path
-          d={
-            'M15 11L13 11L13 9C13 8.73 12.89 8.48 12.7 8.29C12.51 8.1 12.26 8 12 8C11.73 8 11.48 8.1 11.29 8.29C11.1 8.48 11 8.73 11 9L11 11L9 11C8.73 11 8.48 11.1 8.29 11.29C8.1 11.48 8 11.73 8 12C8 12.26 8.1 12.51 8.29 12.7C8.48 12.89 8.73 13 9 13L11 13L11 15C11 15.26 11.1 15.51 11.29 15.7C11.48 15.89 11.73 16 12 16C12.26 16 12.51 15.89 12.7 15.7C12.89 15.51 13 15.26 13 15L13 13L15 13C15.26 13 15.51 12.89 15.7 12.7C15.89 12.51 16 12.26 16 12C16 11.73 15.89 11.48 15.7 11.29C15.51 11.1 15.26 11 15 11Z'
-          }
-          fill={'inherit'}
-          fillOpacity={'1.000000'}
-          fillRule={'nonzero'}
-          id={'Vector'}
-        />
-      </g>
-    </svg>
-  )
-}
