--- src/components/Header/header.tsx
@@ -1,2 +1,0 @@
-import s from '@/components/Header/header.module.scss'
-export const Header = () => <header className={s.header}>header</header>
