--- src/components/Nav/index.ts
@@ -1,1 +1,0 @@
-export * from './Nav'
