+++ src/components/header/index.tsx
@@ -1,0 +1,2 @@
+import s from '@/components/header/header.module.scss'
+export const Header = () => <header className={s.header}>header</header>
