+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export * from './Layout'
+export * from './Layout/baseLayout'
+export * from './Layout/navLayout'
+export * from './headMeta'
+export * from './header'
+export * from './main'
+export * from './nav'
+export * from './pageWrapper'
