+++ src/components/main/main.module.scss
