+++ src/pages/favorites/index.tsx
@@ -1,8 +1,7 @@
-import { getNavLayout } from '@/components/Layout/NavLayout/navLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
+import { GetNavLayout, PageWrapper } from '@/components'
 
 export function Favorites() {
   return <PageWrapper>Favorites</PageWrapper>
 }
-Favorites.getLayout = getNavLayout
+Favorites.getLayout = GetNavLayout
 export default Favorites
