--- src/pages/forgot-password/index.tsx
@@ -1,8 +1,0 @@
-import { getLayout } from '@/components/Layout/BaseLayout/baseLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function ForgotPassword() {
-  return <PageWrapper>ForgotPassword</PageWrapper>
-}
-ForgotPassword.getLayout = getLayout
-export default ForgotPassword
