+++ src/pages/forgotPassword/index.tsx
@@ -1,0 +1,7 @@
+import { GetLayout, PageWrapper } from '@/components'
+
+export function ForgotPassword() {
+  return <PageWrapper>ForgotPassword</PageWrapper>
+}
+ForgotPassword.getLayout = GetLayout
+export default ForgotPassword
