--- src/pages/friend-profile/index.tsx
@@ -1,9 +1,0 @@
-import { getNavLayout } from '@/components/Layout/NavLayout/navLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function FriendProfile() {
-  return <PageWrapper>FriendProfile</PageWrapper>
-}
-
-FriendProfile.getLayout = getNavLayout
-export default FriendProfile
