+++ src/pages/friendProfile/index.tsx
@@ -1,0 +1,8 @@
+import { GetNavLayout, PageWrapper } from '@/components'
+
+export function FriendProfile() {
+  return <PageWrapper>FriendProfile</PageWrapper>
+}
+
+FriendProfile.getLayout = GetNavLayout
+export default FriendProfile
