--- src/pages/general-info/index.tsx
@@ -1,8 +1,0 @@
-import { getNavLayout } from '@/components/Layout/NavLayout/navLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function GeneralInfo() {
-  return <PageWrapper>GeneralInfo</PageWrapper>
-}
-GeneralInfo.getLayout = getNavLayout
-export default GeneralInfo
