+++ src/pages/generalInfo/index.tsx
@@ -1,0 +1,7 @@
+import { GetNavLayout, PageWrapper } from '@/components'
+
+export function GeneralInfo() {
+  return <PageWrapper>GeneralInfo</PageWrapper>
+}
+GeneralInfo.getLayout = GetNavLayout
+export default GeneralInfo
