--- src/pages/privacy-policy/index.tsx
@@ -1,9 +1,0 @@
-import { getLayout } from '@/components/Layout/BaseLayout/baseLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function PrivacyPolicy() {
-  return <PageWrapper>PrivacyPolicy</PageWrapper>
-}
-
-PrivacyPolicy.getLayout = getLayout
-export default PrivacyPolicy
