+++ src/pages/privacyPolicy/index.tsx
@@ -1,0 +1,8 @@
+import { GetLayout, PageWrapper } from '@/components'
+
+export function PrivacyPolicy() {
+  return <PageWrapper>PrivacyPolicy</PageWrapper>
+}
+
+PrivacyPolicy.getLayout = GetLayout
+export default PrivacyPolicy
