--- src/pages/public-page/index.tsx
@@ -1,9 +1,0 @@
-import { getLayout } from '@/components/Layout/BaseLayout/baseLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function PublicPage() {
-  return <PageWrapper>PublicPage</PageWrapper>
-}
-
-PublicPage.getLayout = getLayout
-export default PublicPage
