+++ src/pages/publicPage/index.tsx
@@ -1,0 +1,8 @@
+import { GetLayout, PageWrapper } from '@/components'
+
+export function PublicPage() {
+  return <PageWrapper>PublicPage</PageWrapper>
+}
+
+PublicPage.getLayout = GetLayout
+export default PublicPage
