--- src/pages/sign-in/index.tsx
@@ -1,9 +1,0 @@
-import { getLayout } from '@/components/Layout/BaseLayout/baseLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function SignIn() {
-  return <PageWrapper>SignIn</PageWrapper>
-}
-
-SignIn.getLayout = getLayout
-export default SignIn
