--- src/pages/sign-up/index.tsx
@@ -1,9 +1,0 @@
-import { getLayout } from '@/components/Layout/BaseLayout/baseLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function SignUp() {
-  return <PageWrapper>SignUp</PageWrapper>
-}
-
-SignUp.getLayout = getLayout
-export default SignUp
