+++ src/pages/signIn/index.tsx
@@ -1,0 +1,8 @@
+import { GetLayout, PageWrapper } from '@/components'
+
+export function SignIn() {
+  return <PageWrapper>SignIn</PageWrapper>
+}
+
+SignIn.getLayout = GetLayout
+export default SignIn
