+++ src/pages/signUp/index.tsx
@@ -1,0 +1,8 @@
+import { GetLayout, PageWrapper } from '@/components'
+
+export function SignUp() {
+  return <PageWrapper>SignUp</PageWrapper>
+}
+
+SignUp.getLayout = GetLayout
+export default SignUp
