+++ src/pages/statistics/index.tsx
@@ -1,9 +1,8 @@
-import { getNavLayout } from '@/components/Layout/NavLayout/navLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
+import { GetNavLayout, PageWrapper } from '@/components'
 
 export function Statistics() {
   return <PageWrapper>Statistics</PageWrapper>
 }
 
-Statistics.getLayout = getNavLayout
+Statistics.getLayout = GetNavLayout
 export default Statistics
