--- src/pages/terms-of-service/index.tsx
@@ -1,9 +1,0 @@
-import { getLayout } from '@/components/Layout/BaseLayout/baseLayout'
-import { PageWrapper } from '@/components/PageWrapper/pageWrapper'
-
-export function TermsOfService() {
-  return <PageWrapper>TermsOfService</PageWrapper>
-}
-
-TermsOfService.getLayout = getLayout
-export default TermsOfService
