+++ src/pages/termsOfService/index.tsx
@@ -1,0 +1,8 @@
+import { GetLayout, PageWrapper } from '@/components'
+
+export function TermsOfService() {
+  return <PageWrapper>TermsOfService</PageWrapper>
+}
+
+TermsOfService.getLayout = GetLayout
+export default TermsOfService
