+++ src/pages/userProfile/index.tsx
@@ -1,0 +1,8 @@
+import { GetNavLayout, PageWrapper } from '@/components'
+
+export function UserProfile() {
+  return <PageWrapper>UserProfile</PageWrapper>
+}
+
+UserProfile.getLayout = GetNavLayout
+export default UserProfile
